refactor(core): move string-comment-scan from comments/ to text/
Original location: src/core/src/comments/string-comment-scan.ts
New location:      src/core/src/text/string-comment-scan.ts

Rationale: string-comment-scan.ts is a source-text scanning utility that
tracks string-literal and comment positions simultaneously during raw
character-by-character traversal. Its primary purpose is source-text
navigation (skipping non-code regions), which aligns with the text
utilities in core/src/text/ (alongside line-breaks.ts) rather than
core/src/comments/, which contains comment content processing helpers
(formatting, normalization, doc-comment metadata). The comments/ directory
owns helpers that operate ON comment content; string-comment-scan operates
on raw source text and merely recognizes comment boundaries as part of
its scanning state machine.

All external consumers import through the Core namespace, so no
downstream workspace changes are required. The test file moves to
core/test/text/ to mirror the new source location.

Agent-Logs-Url: https://github.com/SimulatorLife/GMLoop/sessions/51ea52d4-de88-4a3b-ac57-4d57ed8b7267
