+++ src/core/src/comments/index.ts
@@ -1,4 +1,3 @@
 export * from "./comment-utils.js";
 export * from "./doc-comment/index.js";
 export * from "./line-comment/index.js";
-export * from "./string-comment-scan.js";
