+++ src/core/src/index.ts
@@ -59,7 +59,6 @@ export type { StripCommentsTransformOptions } from "./ast/strip-comments-transfo
 export type { GameMakerAstLocation, GameMakerAstNode, LiteralNode, MutableGameMakerAstNode } from "./ast/types.js";
 export type { DocCommentLines, MutableDocCommentLines } from "./comments/comment-utils.js";
 export type { DocCommentNodeMetadata } from "./comments/doc-comment/node-metadata.js";
-export type { StringCommentScanState } from "./comments/string-comment-scan.js";
 export type { GmloopProjectConfig } from "./project-config/gmloop-project-config.js";
 export type { FeatherDiagnostic, FeatherMetadata } from "./resources/feather-metadata.js";
 export type {
@@ -68,5 +67,6 @@ export type {
     DeprecatedIdentifierMetadataEntry,
     DeprecatedIdentifierReplacementKind
 } from "./resources/gml-identifier-loading.js";
+export type { StringCommentScanState } from "./text/string-comment-scan.js";
 export type { AbortSignalLike } from "./utils/abort.js";
 export type { DebouncedFunction } from "./utils/function.js";
