+++ src/core/src/text/index.ts
@@ -1,1 +1,2 @@
 export * from "./line-breaks.js";
+export * from "./string-comment-scan.js";
