+++ src/core/src/text/string-comment-scan.ts
