+++ src/core/test/text/string-comment-scan.test.ts
